requirements.txt
aiomysql==0.1.1
anyio==3.6.2
certifi==2023.5.7
click==8.1.3
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
jose==1.0.0
mysql-connector-python==8.0.33
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
protobuf==3.20.3
pyasn1==0.5.0
pydantic==1.10.8
PyMySQL==1.0.3
pytest==7.3.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
starlette==0.27.0
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.6.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yoyo-migrations==8.2.0
zipp==3.15.0
